fix(wms): 修复质检合格量数据类型和业务逻辑问题
- 将PurchaseOrderDTO中的qcGoodQty字段类型从BigDecimal改为Integer
- 修复质检合格量比较逻辑，统一使用Integer类型进行数值比较
- 新增质检通过时累加合格量的功能实现
- 添加质检撤销时扣减合格量的业务逻辑
- 集成FeignQuery工具类用于查询采购订单详情
- 增加撤销质检时的库存校验防止合格量变为负数

+++ erp-model/erp-model-scm/src/main/java/com/erp/model/scm/dto/PurchaseOrderDTO.java
@@ -2191,6 +2191,6 @@ public class PurchaseOrderDTO implements Serializable {
         /**
          * 质检合格量
          */
-        private BigDecimal qcGoodQty;
+        private Integer qcGoodQty;
     }
 }
